started working os stockExchange

+++ src/Actors/StockOrder.java
@@ -11,9 +11,9 @@ public class StockOrder {
 	final String _type;
 	final int _amount;
 	final String _stockName;
-	final int _price;
+	final Double _price;
 	
-	StockOrder(String type,int amount,String stockName,int price) {
+	StockOrder(String type,int amount,String stockName,Double price) {
 		_type=type;
 		_amount=amount;
 		_stockName=stockName;
